Выносит создание строки продолжительности в функцию
