+++ src/utils/const.js
@@ -33,6 +33,11 @@ const ErrorMessage = {
 
 const DEBOUNCE_DELAY = 100;
 
+const HOURS_IN_DAY = 24;
+const MINUTES_IN_HOUR = 60;
+const NUMBER_OF_CHARACTERS = 2;
+const PAD_STRING = '0';
+
 export {
   FilterType,
   UpdateType,
@@ -40,5 +45,11 @@ export {
   SortType,
   Color,
   ErrorMessage,
-  DEBOUNCE_DELAY
+  DEBOUNCE_DELAY,
+  MenuItem,
+  EVENT_TYPES,
+  HOURS_IN_DAY,
+  MINUTES_IN_HOUR,
+  NUMBER_OF_CHARACTERS,
+  PAD_STRING
 };
